chatbot/requirements.txt
altair==5.2.0
asn1crypto==1.5.1
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
filelock==3.13.1
fonttools==4.47.0
gitdb==4.0.11
GitPython==3.1.40
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
nltk==3.8.1
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
platformdirs==3.11.0
protobuf==4.25.1
pyarrow==14.0.2
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
rich==13.7.0
rpds-py==0.15.2
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
smmap==5.0.1
snowflake-connector-python==3.6.0
sortedcontainers==2.4.0
streamlit==1.29.0
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
tomlkit==0.12.3
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
wordcloud==1.9.3
zipp==3.17.0
